requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
drf-yasg==1.21.7